fixed show vinyl issue; latestVinyl in home

+++ resources/views/user/partials/follow.blade.php
@@ -1,8 +1,15 @@
 <hr>
-@if($user->isFollowedBy(Auth::user()))
-  {!! Form::open([ 'route' => ['unfollow', $user->id], 'method' => 'delete' ]) !!}
-    <button class="btn btn-md btn-default" type="submit"><i class="fa fa-fw fa-minus"></i> Unfollow</button>
-  {!! Form::close() !!}
+@if(Auth::check())
+  @if($user->isFollowedBy(Auth::user()))
+    {!! Form::open([ 'route' => ['unfollow', $user->id], 'method' => 'delete' ]) !!}
+      <button class="btn btn-md btn-default" type="submit"><i class="fa fa-fw fa-minus"></i> Unfollow</button>
+    {!! Form::close() !!}
+  @else
+    {!! Form::open([ 'route' => 'follow' ]) !!}
+      {!! Form::hidden('userIdToFollow', $user->id) !!}
+      <button class="btn btn-md btn-success" type="submit"><i class="fa fa-fw fa-plus"></i> Follow</button>
+    {!! Form::close() !!}
+  @endif
 @else
   {!! Form::open([ 'route' => 'follow' ]) !!}
     {!! Form::hidden('userIdToFollow', $user->id) !!}
